Reverting to commit 3b13f7c. Lint test fails, but tests pass here
