+++ automol/zmat/__init__.py
@@ -109,7 +109,7 @@ from ._ring import ring_distances
 from ._ring import ring_dihedrals
 from ._ring import ring_distances_reasonable
 from ._ring import complete_ring_dihedrals
-from ._ring import checks_with_crest
+
 
 __all__ = [
     # L2
@@ -211,6 +211,5 @@ __all__ = [
     'ring_distances',
     'ring_dihedrals',
     'complete_ring_dihedrals',
-    'ring_distances_reasonable',
-    'checks_with_crest',
+    'ring_distances_reasonable'
 ]
